requirements.txt
slack_bolt
python-novaclient
python-glanceclient
pytest